+++ crates/egui_demo_lib/src/demo/scrolling.rs
@@ -1,4 +1,4 @@
-use egui::{color::*, *};
+use egui::*;
 
 #[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
 #[derive(Clone, Copy, Debug, PartialEq)]